Move sampler serialization
