+++ keras_nlp/samplers/__init__.py
@@ -18,75 +18,8 @@ from keras_nlp.samplers.beam_sampler import BeamSampler
 from keras_nlp.samplers.greedy_sampler import GreedySampler
 from keras_nlp.samplers.random_sampler import RandomSampler
 from keras_nlp.samplers.sampler import Sampler
+from keras_nlp.samplers.serialization import deserialize
+from keras_nlp.samplers.serialization import get
+from keras_nlp.samplers.serialization import serialize
 from keras_nlp.samplers.top_k_sampler import TopKSampler
 from keras_nlp.samplers.top_p_sampler import TopPSampler
-
-
-def serialize(sampler):
-    return keras.utils.serialize_keras_object(sampler)
-
-
-def deserialize(config, custom_objects=None):
-    """Return a `Sampler` object from its config."""
-    all_classes = {
-        "beam": BeamSampler,
-        "greedy": GreedySampler,
-        "random": RandomSampler,
-        "top_k": TopKSampler,
-        "top_p": TopPSampler,
-    }
-    return keras.utils.deserialize_keras_object(
-        config,
-        module_objects=all_classes,
-        custom_objects=custom_objects,
-        printable_module_name="samplers",
-    )
-
-
-def get(identifier):
-    """Retrieve a KerasNLP sampler by the identifier.
-
-    The `identifier` may be the string name of a sampler class or class.
-
-    >>> identifier = 'greedy'
-    >>> sampler = keras_nlp.samplers.get(identifier)
-
-    You can also specify `config` of the sampler to this function by passing
-    dict containing `class_name` and `config` as an identifier. Also note that
-    the `class_name` must map to a `Sampler` class.
-
-    >>> cfg = {'class_name': 'keras_nlp>GreedySampler', 'config': {}}
-    >>> sampler = keras_nlp.samplers.get(cfg)
-
-    In the case that the `identifier` is a class, this method will return a new
-    instance of the class by its constructor.
-
-    Args:
-        identifier: String or dict that contains the sampler name or
-            configurations.
-
-    Returns:
-        Sampler instance base on the input identifier.
-
-    Raises:
-        ValueError: If the input identifier is not a supported type or in a bad
-            format.
-    """
-
-    if identifier is None:
-        return None
-    if isinstance(identifier, dict):
-        return deserialize(identifier)
-    elif isinstance(identifier, str):
-        if not identifier.islower():
-            raise KeyError(
-                "`keras_nlp.samplers.get()` must take a lowercase string "
-                f"identifier, but received: {identifier}."
-            )
-        return deserialize(identifier)
-    elif callable(identifier):
-        return identifier
-    else:
-        raise ValueError(
-            "Could not interpret sampler identifier: " + str(identifier)
-        )
